Moved the dialog to ./dialogs and removed references from voter.js
